Seperate buffering of stem strikes
Stem points buffered independently of the ground points to ensure that
the canopy filter is not too "heavy-handed". It should be noted however
that the problem with the original script is likely to be the
aggregation of stems, based on a lack of a distinct canopy density
signature, and this change will likely make this worse. preliminary
indications are that the compartment stocking predicted is closer to the
value stored in WOODS.

Noted a pattern where-by the nearest ground point to a stem strike was
more than twice the ground-point buffer size, these stem points may
represent dense ground vegetation and should perhaps be removed.